cleanup pools in tests
